Rename file pop-in-campaigns.vue to PopInCampaigns.vue
